+++ packages/postcss-discard-comments/index.js
@@ -38,6 +38,15 @@ module.exports = postcss.plugin('postcss-discard-comments', function (options) {
             if (decl._value && decl._value.raw) {
                 decl._value.raw = replaceComments(decl._value.raw);
             }
+            if (decl._important) {
+                decl._important = replaceComments(decl._important);
+                var b = balanced.matches({
+                    source: decl._important,
+                    open: '/*',
+                    close: '*/'
+                });
+                decl._important = b.length ? decl._important : '!important';
+            }
         });
 
         css.eachRule(function (rule) {